requirements.txt
Django==4.0.5
Pillow==9.3.0
psycopg2-binary==2.9.5
django-cors-headers==4.3.0
djangorestframework==3.14.0
django-environ==0.11.2
django-ses==4.2.0
gunicorn==21.2.0